Version Control Service

+++ corpdir-srv-api/src/main/java/com/github/bordertech/corpdir/api/common/ApiVersionable.java
@@ -7,8 +7,8 @@ package com.github.bordertech.corpdir.api.common;
  */
 public interface ApiVersionable extends ApiKeyIdObject {
 
-	Integer getVersionId();
+	Long getVersionId();
 
-	void setVersionId(final Integer versionId);
+	void setVersionId(final Long versionId);
 
 }
