+++ corpdir-srv-api/src/main/java/com/github/bordertech/corpdir/api/v1/OrgUnitService.java
@@ -1,10 +1,10 @@
 package com.github.bordertech.corpdir.api.v1;
 
 import com.github.bordertech.corpdir.api.response.DataResponse;
+import com.github.bordertech.corpdir.api.service.BasicVersionTreeService;
 import com.github.bordertech.corpdir.api.v1.func.PositionFunctions;
 import com.github.bordertech.corpdir.api.v1.model.OrgUnit;
 import com.github.bordertech.corpdir.api.v1.model.Position;
-import com.github.bordertech.corpdir.api.service.BasicVersionTreeService;
 
 /**
  * Organisation Unit Service Interface.
@@ -16,6 +16,6 @@ public interface OrgUnitService extends BasicVersionTreeService<OrgUnit>, Positi
 
 	DataResponse<Position> getManagerPosition(final String keyId);
 
-	DataResponse<Position> getManagerPosition(final Integer versionId, final String keyId);
+	DataResponse<Position> getManagerPosition(final Long versionId, final String keyId);
 
 }
