+++ corpdir-srv-api/src/main/java/com/github/bordertech/corpdir/api/v1/PositionService.java
@@ -1,11 +1,11 @@
 package com.github.bordertech.corpdir.api.v1;
 
 import com.github.bordertech.corpdir.api.response.DataResponse;
+import com.github.bordertech.corpdir.api.service.BasicVersionTreeService;
+import com.github.bordertech.corpdir.api.v1.func.ContactFunctions;
 import com.github.bordertech.corpdir.api.v1.model.OrgUnit;
 import com.github.bordertech.corpdir.api.v1.model.Position;
 import java.util.List;
-import com.github.bordertech.corpdir.api.v1.func.ContactFunctions;
-import com.github.bordertech.corpdir.api.service.BasicVersionTreeService;
 
 /**
  * Position Service Interface.
@@ -17,6 +17,6 @@ public interface PositionService extends BasicVersionTreeService<Position>, Cont
 
 	DataResponse<List<OrgUnit>> getManages(final String keyId);
 
-	DataResponse<List<OrgUnit>> getManages(final Integer versionId, final String keyId);
+	DataResponse<List<OrgUnit>> getManages(final Long versionId, final String keyId);
 
 }
