+++ corpdir-srv-jpa/src/main/java/com/github/bordertech/corpdir/jpa/common/feature/PersistVersionData.java
@@ -3,7 +3,7 @@ package com.github.bordertech.corpdir.jpa.common.feature;
 import java.util.Set;
 
 /**
- * Object has versioned data.
+ * Object with a Business Key and versioned data.
  *
  * @author jonathan
  * @param <T>
@@ -14,8 +14,8 @@ public interface PersistVersionData<T extends PersistVersionable> extends Persis
 
 	void addDataVersion(final T versionData);
 
-	void removeDataVersion(final Integer versionId);
+	void removeDataVersion(final Long versionId);
 
-	T getDataVersion(final Integer versionId);
+	T getDataVersion(final Long versionId);
 
 }
